+++ packages/core/src/components/axes/ruler.ts
@@ -1,9 +1,9 @@
-import { Selection, pointer } from 'd3'
-import { isEqual } from 'lodash-es'
-import { debounceWithD3MousePosition, getProperty } from '@/tools'
+import { bisectCenter, pointer, selectAll, type Selection } from 'd3'
+import { getProperty } from '@/tools'
 import { Component } from '@/components/component'
 import { DOMUtils } from '@/services/essentials/dom-utils'
 import { CartesianOrientations, Events, RenderTypes } from '@/interfaces/enums'
+import type { ChartTabularData } from '@/interfaces/model'
 
 export type GenericSvgSelection = Selection<SVGGraphicsElement, any, Element, any>
 
@@ -19,11 +19,14 @@ export class Ruler extends Component {
 	renderType = RenderTypes.SVG
 
 	backdrop: GenericSvgSelection
-	elementsToHighlight: GenericSvgSelection
-	pointsWithinLine: {
-		domainValue: number
-		originalData: any
-	}[]
+	elementsToHighlight?: GenericSvgSelection
+	private domainValues?: number[]
+	private pointsByDomain?: Map<number, ChartTabularData>
+	private elementsByDomain?: Map<number, SVGGraphicsElement[]>
+	private activeDomainValue?: number
+	private frame?: number
+	private rulerEvent: CustomEvent
+	private rulerPosition: [number, number]
 	isXGridEnabled = getProperty(this.getOptions(), 'grid', 'x', 'enabled')
 	isYGridEnabled = getProperty(this.getOptions(), 'grid', 'y', 'enabled')
 	// flag for checking whether ruler event listener is added or not
@@ -34,8 +37,16 @@ export class Ruler extends Component {
 	// eslint-disable-next-line @typescript-eslint/no-unused-vars
 	render(animate = false) {
 		const isRulerEnabled = getProperty(this.getOptions(), 'ruler', 'enabled')
-		const alwaysShowRulerTooltip = getProperty(this.getOptions(), 'tooltip', 'alwaysShowRulerTooltip')
+		const alwaysShowRulerTooltip = getProperty(
+			this.getOptions(),
+			'tooltip',
+			'alwaysShowRulerTooltip'
+		)
 		const shouldEnableRuler = isRulerEnabled || alwaysShowRulerTooltip
+		this.domainValues = undefined
+		this.pointsByDomain = undefined
+		this.elementsByDomain = undefined
+		this.activeDomainValue = undefined
 
 		this.drawBackdrop()
 
@@ -49,6 +60,10 @@ export class Ruler extends Component {
 	removeBackdropEventListeners() {
 		this.isEventListenerAdded = false
 		this.backdrop.on('mousemove mouseover mouseout', null)
+		if (this.frame !== undefined) {
+			cancelAnimationFrame(this.frame)
+			this.frame = undefined
+		}
 	}
 
 	formatTooltipData(tooltipData: any) {
@@ -60,143 +75,114 @@ export class Ruler extends Component {
 
 		const orientation: CartesianOrientations = this.services.cartesianScales.getOrientation()
 
-		const displayData = this.model.getDisplayData()
-
 		const rangeScale = this.services.cartesianScales.getRangeScale()
 		const [yScaleEnd, yScaleStart] = rangeScale.range()
 
 		const mouseCoordinate = orientation === CartesianOrientations.HORIZONTAL ? y : x
 		const ruler = DOMUtils.appendOrSelect(svg, 'g.ruler').attr('aria-label', 'ruler')
 		const rulerLine = DOMUtils.appendOrSelect(ruler, 'line.ruler-line')
-		const dataPointElements: GenericSvgSelection = svg.selectAll('[role=graphics-symbol]')
 
-		const pointsWithinLine = displayData
-			.map((d: any) => ({
-				domainValue: this.services.cartesianScales.getDomainValue(d) as number,
-				originalData: d
-			}))
-			.filter((d: any) => pointIsWithinThreshold(d.domainValue, mouseCoordinate))
+		if (!this.pointsByDomain) {
+			this.indexData()
+		}
 
-		if (
-			this.pointsWithinLine &&
-			pointsWithinLine.length === this.pointsWithinLine.length &&
-			pointsWithinLine.map((point: any) => point.domainValue).join() ===
-				this.pointsWithinLine.map(point => point.domainValue).join()
-		) {
-			this.pointsWithinLine = pointsWithinLine
+		const domainValue = this.domainValues[bisectCenter(this.domainValues, mouseCoordinate)]
+		if (domainValue === undefined || !pointIsWithinThreshold(domainValue, mouseCoordinate)) {
+			if (this.elementsToHighlight) this.hideRuler()
+			return
+		}
+
+		if (domainValue === this.activeDomainValue) {
 			return this.services.events.dispatchEvent(Events.Tooltip.MOVE, {
 				mousePosition: [x, y]
 			})
 		}
 
-		this.pointsWithinLine = pointsWithinLine
+		const dataPointsMatchingRulerLine = this.pointsByDomain.get(domainValue)
+		const tooltipData = dataPointsMatchingRulerLine.filter(d => {
+			const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(d)
+			const value = d[rangeIdentifier]
+			return value !== null && value !== undefined
+		})
 
-		/**
-		 * Find matches, reduce is used instead of filter
-		 * to only get elements which belong to the same axis coordinate
-		 */
-		const dataPointsMatchingRulerLine: {
-			domainValue: number
-			originalData: any
-		}[] = this.pointsWithinLine.reduce((accum, currentValue) => {
-			if (accum.length === 0) {
-				accum.push(currentValue)
-				return accum
-			}
+		if (!this.elementsByDomain) {
+			this.indexElements()
+		}
+		const elementsToHighlight = selectAll(
+			this.elementsByDomain.get(domainValue) ?? []
+		) as GenericSvgSelection
 
-			// store the first element of the accumulator array to compare it with current element being processed
-			const sampleAccumValue = accum[0].domainValue
+		this.elementsToHighlight?.dispatch('mouseout')
+		elementsToHighlight.dispatch('mouseover')
+		this.elementsToHighlight = elementsToHighlight
+		this.activeDomainValue = domainValue
 
-			const distanceToCurrentValue = Math.abs(mouseCoordinate - currentValue.domainValue)
-			const distanceToAccumValue = Math.abs(mouseCoordinate - sampleAccumValue)
+		this.services.events.dispatchEvent(Events.Tooltip.SHOW, {
+			event,
+			mousePosition: [x, y],
+			hoveredElement: rulerLine,
+			data: this.formatTooltipData(tooltipData)
+		})
 
-			if (distanceToCurrentValue > distanceToAccumValue) {
-				// if distance with current value is bigger than already existing value in the accumulator, skip current iteration
-				return accum
-			} else if (distanceToCurrentValue < distanceToAccumValue) {
-				// currentValue data point is closer to mouse inside the threshold area, so reinstantiate array
-				accum = [currentValue]
-			} else {
-				// currentValue is equal to already stored values, which means there's another match on the same coordinate
-				accum.push(currentValue)
-			}
+		ruler.attr('opacity', 1)
 
-			return accum
-		}, [])
-
-		// some data point match
-		if (dataPointsMatchingRulerLine.length > 0) {
-			const tooltipData = dataPointsMatchingRulerLine
-				.map((d: any) => d.originalData)
-				.filter((d: any) => {
-					const rangeIdentifier = this.services.cartesianScales.getRangeIdentifier(d)
-					const value = d[rangeIdentifier]
-					return value !== null && value !== undefined
-				})
-
-			// get elements on which we should trigger mouse events
-			const domainValuesMatchingRulerLine = dataPointsMatchingRulerLine.map(
-				(d: any) => d.domainValue
-			)
-			const elementsToHighlight = dataPointElements.filter((d: any) => {
-				const domainValue = this.services.cartesianScales.getDomainValue(d) as number
-				return domainValuesMatchingRulerLine.includes(domainValue)
-			})
-
-			/** if we pass from a trigger area to another one
-			 * mouseout on previous elements won't get dispatched
-			 * so we need to do it manually
-			 */
-			if (
-				this.elementsToHighlight &&
-				this.elementsToHighlight.size() > 0 &&
-				!isEqual(this.elementsToHighlight, elementsToHighlight)
-			) {
-				this.hideRuler()
-			}
-
-			elementsToHighlight.dispatch('mouseover')
-
-			// set current hovered elements
-			this.elementsToHighlight = elementsToHighlight
-
-			this.services.events.dispatchEvent(Events.Tooltip.SHOW, {
-				event,
-				mousePosition: [x, y],
-				hoveredElement: rulerLine,
-				data: this.formatTooltipData(tooltipData)
-			})
-
-			ruler.attr('opacity', 1)
-
-			// line snaps to matching point
-			const sampleMatch = dataPointsMatchingRulerLine[0]
-			if (orientation === 'horizontal') {
-				rulerLine
-					.attr('x1', yScaleStart)
-					.attr('x2', yScaleEnd)
-					.attr('y1', sampleMatch.domainValue)
-					.attr('y2', sampleMatch.domainValue)
-			} else {
-				rulerLine
-					.attr('y1', yScaleStart)
-					.attr('y2', yScaleEnd)
-					.attr('x1', sampleMatch.domainValue)
-					.attr('x2', sampleMatch.domainValue)
-			}
+		if (orientation === 'horizontal') {
+			rulerLine
+				.attr('x1', yScaleStart)
+				.attr('x2', yScaleEnd)
+				.attr('y1', domainValue)
+				.attr('y2', domainValue)
 		} else {
-			this.hideRuler()
+			rulerLine
+				.attr('y1', yScaleStart)
+				.attr('y2', yScaleEnd)
+				.attr('x1', domainValue)
+				.attr('x2', domainValue)
 		}
 	}
 
 	hideRuler() {
 		const svg = this.parent
 		const ruler = DOMUtils.appendOrSelect(svg, 'g.ruler')
-		const dataPointElements = svg.selectAll('[role=graphics-symbol]')
 
-		dataPointElements.dispatch('mouseout')
+		this.elementsToHighlight?.dispatch('mouseout')
+		this.elementsToHighlight = undefined
+		this.activeDomainValue = undefined
 		this.services.events.dispatchEvent(Events.Tooltip.HIDE)
 		ruler.attr('opacity', 0)
+	}
+
+	private indexData() {
+		const pointsByDomain = new Map<number, ChartTabularData>()
+		this.model.getDisplayData().forEach(originalData => {
+			const domainValue = this.services.cartesianScales.getDomainValue(originalData) as number
+			if (!Number.isFinite(domainValue)) return
+			const points = pointsByDomain.get(domainValue)
+			if (points) {
+				points.push(originalData)
+			} else {
+				pointsByDomain.set(domainValue, [originalData])
+			}
+		})
+
+		this.pointsByDomain = pointsByDomain
+		this.domainValues = [...pointsByDomain.keys()].sort((a, b) => a - b)
+	}
+
+	private indexElements() {
+		const elementsByDomain = new Map<number, SVGGraphicsElement[]>()
+		this.parent
+			.selectAll<SVGGraphicsElement, ChartTabularData[number]>('[role=graphics-symbol]')
+			.each((d, i, nodes) => {
+				const domainValue = this.services.cartesianScales.getDomainValue(d) as number
+				const elements = elementsByDomain.get(domainValue)
+				if (elements) {
+					elements.push(nodes[i])
+				} else {
+					elementsByDomain.set(domainValue, [nodes[i]])
+				}
+			})
+		this.elementsByDomain = elementsByDomain
 	}
 
 	/**
@@ -204,36 +190,24 @@ export class Ruler extends Component {
 	 */
 	addBackdropEventListeners() {
 		this.isEventListenerAdded = true
-
-		const self = this
-
-		const holder = this.services.domUtils.getHolder()
-
-		const displayData = this.model.getDisplayData()
-
-		let mouseMoveCallback = function (event: CustomEvent) {
-			const pos = pointer(event, self.parent.node())
-
-			self.showRuler(event, pos)
-		}
-
-		// Debounce mouseMoveCallback if there are more than 100 datapoints
-		if (displayData.length > 100) {
-			const debounceThreshold = (displayData.length % 50) * 12.5
-
-			mouseMoveCallback = debounceWithD3MousePosition(
-				function (event: CustomEvent) {
-					const { mousePosition } = this
-					self.showRuler(event, mousePosition)
-				},
-				debounceThreshold,
-				holder
-			)
-		}
-
 		this.backdrop
-			.on('mousemove mouseover', mouseMoveCallback)
-			.on('mouseout', this.hideRuler.bind(this))
+			.on('mousemove mouseover', (event: CustomEvent) => {
+				this.rulerEvent = event
+				this.rulerPosition = pointer(event, this.parent.node())
+				if (this.frame === undefined) {
+					this.frame = requestAnimationFrame(() => {
+						this.frame = undefined
+						this.showRuler(this.rulerEvent, this.rulerPosition)
+					})
+				}
+			})
+			.on('mouseout', () => {
+				if (this.frame !== undefined) {
+					cancelAnimationFrame(this.frame)
+					this.frame = undefined
+				}
+				this.hideRuler()
+			})
 	}
 
 	drawBackdrop() {